Consolidate service providers
