+++ app/Http/Controllers/Auth/VerifyEmailController.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers\Auth;
 
+use App\Providers\AppServiceProvider;
 use App\Http\Controllers\Controller;
 use App\Providers\RouteServiceProvider;
 use Illuminate\Auth\Events\Verified;
@@ -26,7 +27,7 @@ class VerifyEmailController extends Controller
         // @phpstan-ignore-next-line
         if ($request->user()->hasVerifiedEmail()) {
             return redirect()
-                ->intended(RouteServiceProvider::HOME . '?verified=1');
+                ->intended(AppServiceProvider::HOME . '?verified=1');
         }
 
         /** @var MustVerifyEmail */
@@ -35,6 +36,6 @@ class VerifyEmailController extends Controller
             event(new Verified($user));
         }
 
-        return redirect()->intended(RouteServiceProvider::HOME . '?verified=1');
+        return redirect()->intended(AppServiceProvider::HOME . '?verified=1');
     }
 }
