--- app/Providers/AuthServiceProvider.php
@@ -1,34 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Providers;
-
-use App\Models\Campaign;
-use App\Models\Event;
-use App\Policies\CampaignPolicy;
-use App\Policies\EventPolicy;
-use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
-
-/**
- * @psalm-suppress UnusedClass
- */
-class AuthServiceProvider extends ServiceProvider
-{
-    /**
-     * The policy mappings for the application.
-     * @var array<class-string, class-string>
-     */
-    protected $policies = [
-        Campaign::class => CampaignPolicy::class,
-        Event::class => EventPolicy::class,
-    ];
-
-    /**
-     * Register any authentication / authorization services.
-     */
-    public function boot(): void
-    {
-        $this->registerPolicies();
-    }
-}
