+++ tests/Feature/Http/Controllers/Auth/RegisteredUserControllerTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Tests\Feature\Http\Controllers\Auth;
 
+use App\Providers\AppServiceProvider;
 use App\Models\Campaign;
 use App\Models\CampaignInvitation;
 use App\Providers\RouteServiceProvider;
@@ -31,7 +32,7 @@ final class RegisteredUserControllerTest extends TestCase
             'password' => 'password',
             'password_confirmation' => 'password',
         ])
-            ->assertRedirect(RouteServiceProvider::HOME);
+            ->assertRedirect(AppServiceProvider::HOME);
         self::assertAuthenticated();
     }
